+++ 2025/src/main/java/dpr/aoc2025/Day10.java
@@ -18,11 +18,11 @@ class Day10 implements Day {
         return 10;
     }
 
-    int part1(List<String> lines) {
+    long part1(List<String> lines) {
         return 0;
     }
 
-    int part2(List<String> lines) {
+    long part2(List<String> lines) {
         return 0;
     }
